requirements.txt
asgiref==3.7.2
Django==4.2.1
django-environ==0.10.0
django-smtp-ssl==1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
